requirements.txt
docker==7.1.0
Flask==3.1.0
kafka-python-ng==2.2.3
loguru==0.7.3
websockets==15.0
